+++ kwcoco/cli/coco_toydata.py
@@ -10,11 +10,26 @@ class CocoToyDataCLI(object):
         """
         default = {
             'key': scfg.Value('shapes8', help='special demodata code', position=1),
+
             'dst': scfg.Value('test.mscoco.json', help='output path', position=2),
+
+            'image_root': scfg.Value(None, help='path to output the images to')
+
+            'bundle_dpath': scfg.Value(None, help=ub.paragraph(
+                '''
+                If specified, overwrites the `dst` and `img_root` parameters
+                and creates a bundled dataset.
+
+                '''),
         }
         epilog = """
         Example Usage:
             kwcoco toydata --key=shapes8 --dst=toydata.mscoco.json
+
+            kwcoco toydata \
+                --key=shapes8 \
+                --dst=./shapes8.kwcoco/dataset.kwcoco.json
+                --image_root=./shapes8.kwcoco/.assets/images
 
         TODO:
             - [ ] allow specification of images directory